requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-rest-auth==6.0.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflection==0.5.1
kombu==5.4.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.5
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
